move ascii_art to gui package to align with Single Responsibility Principle (SRP) and update README structure
