+++ gui/ascii_art.py
